Update to rust nightly, needs review

+++ src/common/settings.rs
@@ -1,2 +1,1 @@
-pub const MAX_MANIFOLD_POINTS: uint = 2;
-//static MAX_POLYGON_VERTICES: uint = 8;
+pub const MAX_MANIFOLD_POINTS: usize = 2;
